requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Web Scraping
playwright==1.40.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Future Features (uncomment when needed)
# sqlalchemy==2.0.23
# alembic==1.12.1
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# redis==5.0.1
# celery==5.3.4